@@ -56,7 +56,7 @@
     "@babel/plugin-syntax-import-meta": "^7.8.3",
     "@babel/plugin-transform-react-constant-elements": "^7.8.3",
     "@babel/plugin-transform-react-inline-elements": "^7.8.3",
-    "@babel/preset-env": "^7.8.3",
+    "@babel/preset-env": "^7.8.4",
     "@babel/preset-react": "^7.8.3",
     "babel-core": "^7.0.0-bridge.0",
     "babel-eslint": "^9.0.0",
@@ -69,9 +69,9 @@
     "enzyme-adapter-react-16": "^1.15.2",
     "eslint": "^4.19.1",
     "eslint-config-airbnb": "^16.1.0",
-    "eslint-plugin-import": "^2.20.0",
+    "eslint-plugin-import": "^2.20.1",
     "eslint-plugin-jsx-a11y": "^6.2.3",
-    "eslint-plugin-react": "^7.18.0",
+    "eslint-plugin-react": "^7.18.3",
     "file-loader": "^2.0.0",
     "jest": "^24.9.0",
     "mini-css-extract-plugin": "^0.8.2",
@@ -87,16 +87,17 @@
     "sass-loader": "^6.0.7",
     "style-loader": "^1.1.3",
     "terser-webpack-plugin": "^1.4.3",
-    "webpack-cli": "^3.3.10",
+    "webpack-cli": "^3.3.11",
     "webpack-manifest-plugin": "^2.2.0",
     "webpack-node-externals": "^1.7.2"
   },
   "engines": {
-    "node": "11.15.0",
-    "npm": "6.7.0"
+    "node": "12.16.1",
+    "npm": "6.13.4"
   },
   "dependencies": {
-    "@babel/core": "^7.8.3",
+    "@auth0/s3": "^1.0.0",
+    "@babel/core": "^7.8.4",
     "@babel/polyfill": "^7.8.3",
     "@babel/register": "^7.8.3",
     "archiver": "^1.1.0",
@@ -108,7 +109,7 @@
     "bson-objectid": "^1.3.0",
     "classnames": "^2.2.6",
     "clipboard": "^1.7.1",
-    "codemirror": "^5.51.0",
+    "codemirror": "^5.52.0",
     "connect-mongo": "^1.3.2",
     "console-feed": "^2.8.11",
     "cookie-parser": "^1.4.3",
@@ -133,15 +134,17 @@
     "jshint": "^2.11.0",
     "lodash": "^4.17.15",
     "loop-protect": "github:catarak/loop-protect",
+    "mime-types": "^2.1.26",
     "mjml": "^3.3.2",
     "mockingoose": "^2.13.2",
-    "mongoose": "^5.8.10",
+    "mongoose": "^5.9.2",
     "node-uuid": "^1.4.7",
     "nodemailer": "^2.6.4",
     "nodemailer-mailgun-transport": "^1.4.0",
     "passport": "^0.3.2",
     "passport-github": "^1.1.0",
     "passport-google-oauth20": "^1.0.0",
+    "passport-http": "^0.3.0",
     "passport-local": "^1.0.0",
     "pretty-bytes": "^3.0.1",
     "primer-tooltips": "^1.5.11",
@@ -157,15 +160,15 @@
     "react-split-pane": "^0.1.89",
     "react-tabs": "^2.3.1",
     "redux": "^3.7.2",
+    "redux-auth-wrapper": "^2.1.0",
     "redux-devtools": "^3.4.2",
     "redux-devtools-dock-monitor": "^1.1.3",
     "redux-devtools-log-monitor": "^1.4.0",
     "redux-form": "^5.3.3",
     "redux-thunk": "^2.3.0",
-    "request": "^2.88.0",
+    "request": "^2.88.2",
     "request-promise": "^4.2.5",
     "reselect": "^4.0.0",
-    "s3": "^4.4.0",
     "s3-policy": "^0.2.0",
     "sass-extract": "^2.1.0",
     "sass-extract-js": "^0.4.0",
@@ -176,7 +179,7 @@
     "slugify": "^1.3.6",
     "srcdoc-polyfill": "^0.2.0",
     "url": "^0.11.0",
-    "webpack": "^4.41.5",
+    "webpack": "^4.41.6",
     "webpack-dev-middleware": "^2.0.6",
     "webpack-hot-middleware": "^2.24.3",
     "xhr": "^2.5.0"
