@@ -1,55 +1,33 @@
 
 import PropTypes from 'prop-types';
 import React from 'react';
-import { Link, browserHistory } from 'react-router';
+import { Link } from 'react-router';
 import classNames from 'classnames';
-import InlineSVG from 'react-inlinesvg';
 import { bindActionCreators } from 'redux';
 import { reduxForm } from 'redux-form';
 import { Helmet } from 'react-helmet';
 import * as UserActions from '../actions';
 import ResetPasswordForm from '../components/ResetPasswordForm';
 import { validateResetPassword } from '../../../utils/reduxFormUtils';
+import Nav from '../../../components/Nav';
 
-const exitUrl = require('../../../images/exit.svg');
-const logoUrl = require('../../../images/p5js-logo.svg');
-
-class ResetPasswordView extends React.Component {
-  constructor(props) {
-    super(props);
-    this.gotoHomePage = this.gotoHomePage.bind(this);
-  }
-
-  componentWillMount() {
-    this.props.resetPasswordReset();
-  }
-
-  gotoHomePage() {
-    browserHistory.push('/');
-  }
-
-  render() {
-    const resetPasswordClass = classNames({
-      'reset-password': true,
-      'reset-password--submitted': this.props.user.resetPasswordInitiate,
-      'form-container': true
-    });
-    return (
+function ResetPasswordView(props) {
+  const resetPasswordClass = classNames({
+    'reset-password': true,
+    'reset-password--submitted': props.user.resetPasswordInitiate,
+    'form-container': true,
+    'user': true
+  });
+  return (
+    <div className="reset-password-container">
+      <Nav layout="dashboard" />
       <div className={resetPasswordClass}>
         <Helmet>
           <title>p5.js Web Editor | Reset Password</title>
         </Helmet>
-        <div className="form-container__header">
-          <button className="form-container__logo-button" onClick={this.gotoHomePage}>
-            <InlineSVG src={logoUrl} alt="p5js Logo" />
-          </button>
-          <button className="form-container__exit-button" onClick={this.gotoHomePage}>
-            <InlineSVG src={exitUrl} alt="Close ResetPassword Page" />
-          </button>
-        </div>
         <div className="form-container__content">
           <h2 className="form-container__title">Reset Your Password</h2>
-          <ResetPasswordForm {...this.props} />
+          <ResetPasswordForm {...props} />
           <p className="reset-password__submitted">
             Your password reset email should arrive shortly. If you don&apos;t see it, check
             in your spam folder as sometimes it can end up there.
@@ -61,8 +39,8 @@ class ResetPasswordView extends React.Component {
           </p>
         </div>
       </div>
-    );
-  }
+    </div>
+  );
 }
 
 ResetPasswordView.propTypes = {
