@@ -1,13 +1,14 @@
 import slugify from 'slugify';
 import friendlyWords from 'friendly-words';
+import lodash from 'lodash';
+
+import passport from 'passport';
+import GitHubStrategy from 'passport-github';
+import LocalStrategy from 'passport-local';
+import GoogleStrategy from 'passport-google-oauth20';
+import { BasicStrategy } from 'passport-http';
+
 import User from '../models/user';
-
-const lodash = require('lodash');
-const passport = require('passport');
-const GitHubStrategy = require('passport-github').Strategy;
-const LocalStrategy = require('passport-local').Strategy;
-const GoogleStrategy = require('passport-google-oauth20').Strategy;
-
 
 passport.serializeUser((user, done) => {
   done(null, user.id);
@@ -36,6 +37,24 @@ passport.use(new LocalStrategy({ usernameField: 'email' }, (email, password, don
       });
     })
     .catch(err => done(null, false, { msg: err }));
+}));
+
+/**
+ * Authentificate using Basic Auth (Username + Api Key)
+ */
+passport.use(new BasicStrategy((userid, key, done) => {
+  User.findOne({ username: userid }, (err, user) => { // eslint-disable-line consistent-return
+    if (err) { return done(err); }
+    if (!user) { return done(null, false); }
+    user.findMatchingKey(key, (innerErr, isMatch, keyDocument) => {
+      if (isMatch) {
+        keyDocument.lastUsedAt = Date.now();
+        user.save();
+        return done(null, user);
+      }
+      return done(null, false, { msg: 'Invalid username or API key' });
+    });
+  });
 }));
 
 /*
