@@ -1,5 +1,7 @@
 import passport from 'passport';
 
+import { userResponse } from './user.controller';
+
 export function createSession(req, res, next) {
   passport.authenticate('local', (err, user) => { // eslint-disable-line consistent-return
     if (err) { return next(err); }
@@ -9,26 +11,14 @@ export function createSession(req, res, next) {
 
     req.logIn(user, (innerErr) => {
       if (innerErr) { return next(innerErr); }
-      return res.json({
-        email: req.user.email,
-        username: req.user.username,
-        preferences: req.user.preferences,
-        verified: req.user.verified,
-        id: req.user._id
-      });
+      return res.json(userResponse(req.user));
     });
   })(req, res, next);
 }
 
 export function getSession(req, res) {
   if (req.user) {
-    return res.json({
-      email: req.user.email,
-      username: req.user.username,
-      preferences: req.user.preferences,
-      verified: req.user.verified,
-      id: req.user._id
-    });
+    return res.json(userResponse(req.user));
   }
   return res.status(404).send({ message: 'Session does not exist' });
 }
