@@ -13,7 +13,7 @@ mongoose.connection.on('error', () => {
 import Project from '../models/project';
 import User from '../models/user';
 
-import s3 from 's3';
+import s3 from '@auth0/s3';
 
 let client = s3.createClient({
   maxAsyncS3: 20,
