@@ -1,3 +1,7 @@
 require('@babel/register');
 require('@babel/polyfill');
-require('./truncate');
+const path = require('path');
+require('dotenv').config({ path: path.resolve('.env') });
+require('./populateTotalSize');
+// require('./moveBucket');
+// require('./truncate');
