@@ -49,8 +49,43 @@ projectSchema.set('toJSON', {
 
 projectSchema.pre('save', function generateSlug(next) {
   const project = this;
-  project.slug = slugify(project.name, '_');
+
+  if (!project.slug) {
+    project.slug = slugify(project.name, '_');
+  }
+
   return next();
 });
+
+/**
+ * Check if slug is unique for this user's projects
+ */
+projectSchema.methods.isSlugUnique = async function isSlugUnique(cb) {
+  const project = this;
+  const hasCallback = typeof cb === 'function';
+
+  try {
+    const docsWithSlug = await project.model('Project')
+      .find({ user: project.user, slug: project.slug }, '_id')
+      .exec();
+
+    const result = {
+      isUnique: docsWithSlug.length === 0,
+      conflictingIds: docsWithSlug.map(d => d._id) || []
+    };
+
+    if (hasCallback) {
+      cb(null, result);
+    }
+
+    return result;
+  } catch (err) {
+    if (hasCallback) {
+      cb(err, null);
+    }
+
+    throw err;
+  }
+};
 
 export default mongoose.model('Project', projectSchema);
