@@ -6,7 +6,7 @@ const router = new Router();
 
 router.post('/S3/sign', isAuthenticated, AWSController.signS3);
 router.post('/S3/copy', isAuthenticated, AWSController.copyObjectInS3);
-router.delete('/S3/:object_key', isAuthenticated, AWSController.deleteObjectFromS3);
-router.get('/S3/objects', AWSController.listObjectsInS3ForUser);
+router.delete('/S3/:userId?/:objectKey', isAuthenticated, AWSController.deleteObjectFromS3);
+router.get('/S3/objects', AWSController.listObjectsInS3ForUserRequestHandler);
 
 export default router;
